modify

+++ lightning_attn_cute/test.py
@@ -159,8 +159,8 @@ def torch_compute_kv(k, v, BLOCK = 64):
     for i in range(NUM_BLOCK):
         si = i * BLOCK
         ei = min(si + BLOCK, N)
-        ki = k[:, :, si:ei].contiguous().to(torch.float32)
-        vi = v[:, :, si:ei].contiguous().to(torch.float32)
+        ki = k[:, :, si:ei].contiguous().to(torch.float16)
+        vi = v[:, :, si:ei].contiguous().to(torch.float16)
 
         new_kv = torch.matmul(ki.transpose(-1, -2), vi)
         print(f"new_kv type: {new_kv.dtype}")
